fix: render on document.body.className changes
In https://github.com/gregorias/anki-code-highlighter/issues/86, we’ve
discovered that Simple Cloze Overlapper renders a blank screen. I’ve
discovered that’s due to the plugin’s script not working with an
up-to-date class list
(https://forums.ankiweb.net/t/card-card-classes-only-injected-separately-now/27387)

This commit adds an observer to className to address the issue.